Closes #876 https://github.com/winjs/winjs/issues/876
The issue is that the menu Layout was animating the toolbar when switching between 'minimal' and 'compact' closedDisplayModes, However, that transition does not require a Layout animation, only an AppBar animation, which the AppBar does for all Layouts.
The fix was to adjust the Menu layout to only run its animations whenever we are changing to or from "shown" position.